chore:(@biblical/web) hooks query 폴더로 분리
